Fixed the asteroid bug by setting the starting heading to 0. I don't understand why it didn't work before with large headings which worries me.

+++ blasteroids.h
@@ -26,7 +26,7 @@ typedef struct Object {
 	float heading;
 	float acceleration;
 	float size;
-	int max_speed;
+	float max_speed;
 	int state;
 	ALLEGRO_COLOR color;
 	struct Object* next;
